device-identifier/requirements.txt
grpcio==1.48.0
grpcio-reflection==1.45.0
grpcio-tools==1.16.1
pyinstaller==5.1
click==8.1.3
pyshark==0.5.3
pyfiglet==0.8.post1
loguru==0.6.0
influxdb
influxdb-client==1.30.0
pyinstaller==5.1
pandas==1.4.3